requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.5
fire==0.4.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
robottest==0.0.1
six==1.16.0
termcolor==1.1.0
tomli==2.0.1
